pyproject.toml
[project]
name = "leoroar"
version = "0.1.0"
description = "leoroar using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai-tools>=1.10.1",
    "crewai[google-genai,tools]==1.10.1",
    "modal>=1.3.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.55.0",
    "telethon>=1.42.0",
]

[project.scripts]
leoroar = "leoroar.main:run"
run_crew = "leoroar.main:run"
train = "leoroar.main:train"
replay = "leoroar.main:replay"
test = "leoroar.main:test"
run_with_trigger = "leoroar.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
